+++ src/o365/aad/commands/o365group/report/o365group-report-activitydetail.ts
@@ -1,7 +1,7 @@
-import commands from '../../commands';
-import DateAndPeriodBasedReport from '../../../base/DateAndPeriodBasedReport';
+import commands from '../../../commands';
+import DateAndPeriodBasedReport from '../../../../base/DateAndPeriodBasedReport';
 
-const vorpal: Vorpal = require('../../../../vorpal-init');
+const vorpal: Vorpal = require('../../../../../vorpal-init');
 
 class O365GroupReportActivityDetailCommand extends DateAndPeriodBasedReport {
   public get name(): string {
@@ -29,11 +29,11 @@ class O365GroupReportActivityDetailCommand extends DateAndPeriodBasedReport {
 
     Get details about Office 365 Groups activity by group for the last week
     and exports the report data in the specified path in text format
-      ${commands.O365GROUP_REPORT_ACTIVITYDETAIL} --period D7 --output text --outputFile './o365groupactivitydetail.txt'
+      ${commands.O365GROUP_REPORT_ACTIVITYDETAIL} --period D7 --output text --outputFile o365groupactivitydetail.txt
 
     Get details about Office 365 Groups activity by group for the last week
     and exports the report data in the specified path in json format
-      ${commands.O365GROUP_REPORT_ACTIVITYDETAIL} --period D7 --output json --outputFile './o365groupactivitydetail.json'
+      ${commands.O365GROUP_REPORT_ACTIVITYDETAIL} --period D7 --output json --outputFile o365groupactivitydetail.json
 `);
   }
 }