+++ src/o365/aad/commands/o365group/report/o365group-report-activitygroupcounts.spec.ts
@@ -1,14 +1,14 @@
-import commands from '../../commands';
+import commands from '../../../commands';
+import Command from '../../../../../Command';
 import * as sinon from 'sinon';
-import appInsights from '../../../../appInsights';
-import auth from '../../../../Auth';
-const command: Command = require('./o365group-report-activitydetail');
+import appInsights from '../../../../../appInsights';
+import auth from '../../../../../Auth';
+const command: Command = require('./o365group-report-activitygroupcounts');
 import * as assert from 'assert';
-import Utils from '../../../../Utils';
-import request from '../../../../request';
-import Command from '../../../../Command';
+import Utils from '../../../../../Utils';
+import request from '../../../../../request';
 
-describe(commands.O365GROUP_REPORT_ACTIVITYDETAIL, () => {
+describe(commands.O365GROUP_REPORT_ACTIVITYGROUPCOUNTS, () => {
   let vorpal: Vorpal;
   let log: string[];
   let cmdInstance: any;
@@ -20,7 +20,7 @@ describe(commands.O365GROUP_REPORT_ACTIVITYDETAIL, () => {
   });
 
   beforeEach(() => {
-    vorpal = require('../../../../vorpal-init');
+    vorpal = require('../../../../../vorpal-init');
     log = [];
     cmdInstance = {
       commandWrapper: {
@@ -50,20 +50,20 @@ describe(commands.O365GROUP_REPORT_ACTIVITYDETAIL, () => {
   });
 
   it('has correct name', () => {
-    assert.equal(command.name.startsWith(commands.O365GROUP_REPORT_ACTIVITYDETAIL), true);
+    assert.equal(command.name.startsWith(commands.O365GROUP_REPORT_ACTIVITYGROUPCOUNTS), true);
   });
 
   it('has a description', () => {
     assert.notEqual(command.description, null);
   });
 
-  it('gets details about Office 365 Groups activity by group for the given period', (done) => {
+  it('gets the daily total number of groups and how many of them were active based on activities for the given period', (done) => {
     const requestStub: sinon.SinonStub = sinon.stub(request, 'get').callsFake((opts) => {
-      if (opts.url === `https://graph.microsoft.com/v1.0/reports/getOffice365GroupsActivityDetail(period='D7')`) {
+      if (opts.url === `https://graph.microsoft.com/v1.0/reports/getOffice365GroupsActivityGroupCounts(period='D7')`) {
         return Promise.resolve(`
-        Report Refresh Date,Group Display Name,Is Deleted,Owner Principal Name,Last Activity Date,Group Type,Member Count,External Member Count,Exchange Received Email Count,SharePoint Active File Count,Yammer Posted Message Count,Yammer Read Message Count,Yammer Liked Message Count,Exchange Mailbox Total Item Count,Exchange Mailbox Storage Used (Byte),SharePoint Total File Count,SharePoint Site Storage Used (Byte),Group Id,Report Period
-        2019-10-01,Pavithra Library,False,user1@sharepointrider.onmicrosoft.com,,Private,7,2,,,,,,430,4757931,0,1450329,01c48e08-ff4a-4d47-bb42-947581d1b3fe,7
-        2019-10-01,D.Marketing,True,user2@sharepointrider.onmicrosoft.com,2019-05-30,Private,4,0,,,,,,413,3882649,4,1596856,02826124-adbe-4d57-8ccb-a2b5647cad14,7
+        Report Refresh Date,Total,Active,Report Date,Report Period
+        2019-10-14,217,0,2019-10-14,7
+        2019-10-14,217,0,2019-10-13,7
         `);
       }
 
@@ -72,7 +72,7 @@ describe(commands.O365GROUP_REPORT_ACTIVITYDETAIL, () => {
 
     cmdInstance.action({ options: { debug: false, period: 'D7' } }, () => {
       try {
-        assert.equal(requestStub.lastCall.args[0].url, "https://graph.microsoft.com/v1.0/reports/getOffice365GroupsActivityDetail(period='D7')");
+        assert.equal(requestStub.lastCall.args[0].url, "https://graph.microsoft.com/v1.0/reports/getOffice365GroupsActivityGroupCounts(period='D7')");
         assert.equal(requestStub.lastCall.args[0].headers["accept"], 'application/json;odata.metadata=none');
         assert.equal(requestStub.lastCall.args[0].json, true);
         done();
@@ -92,7 +92,7 @@ describe(commands.O365GROUP_REPORT_ACTIVITYDETAIL, () => {
     const find = sinon.stub(vorpal, 'find').callsFake(() => cmd);
     cmd.help = command.help();
     cmd.help({}, () => { });
-    assert(find.calledWith(commands.O365GROUP_REPORT_ACTIVITYDETAIL));
+    assert(find.calledWith(commands.O365GROUP_REPORT_ACTIVITYGROUPCOUNTS));
   });
 
   it('has help with examples', () => {